refactor(oxlint,oxfmt): Fix up `apps/shared` usage (#22457)
Final clean up.

- Share `isObject()` function
- Clean up directory structure
- Use `Error.cause` for TS hint error
