+++ apps/shared/src-js/index.ts
@@ -1,2 +1,2 @@
-export { importJsConfig } from "./js_config.ts";
-export { loadViteConfigField } from "./vite_plus.ts";
+export { importJsConfig } from "./js_config/index.ts";
+export { loadViteConfigField } from "./vite_plus_config.ts";
