+++ apps/shared/src-js/utils.ts
@@ -1,0 +1,2 @@
+export const isObject = (v: unknown): v is object =>
+  typeof v === "object" && v !== null && !Array.isArray(v);
